+++ metro-extensions/contribute/impl-compiler-plugin/src/main/kotlin/software/ralf/app/platform/metro/compiler/ClassIds.kt
@@ -5,8 +5,6 @@ import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.name.Name
 
 internal object ClassIds {
-  val ANY = ClassId(FqName("kotlin"), Name.identifier("Any"))
-
   val APP_SCOPE = ClassId(FqName("dev.zacsweers.metro"), Name.identifier("AppScope"))
 
   val BASE_MODEL =