+++ metro-extensions/contribute/impl-compiler-plugin/src/main/kotlin/software/ralf/app/platform/metro/compiler/renderer/ContributesRendererMetroExtension.kt
@@ -66,8 +66,7 @@ public class ContributesRendererMetroExtension(private val session: FirSession) 
       session: FirSession,
       options: MetroOptions,
       compatContext: CompatContext,
-    ): MetroContributionExtension? {
-      if (options.generateClassesInIr) return null
+    ): MetroContributionExtension {
       return ContributesRendererMetroExtension(session)
     }
   }